+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 ### Added
 
 - Add `HuangYi\Shadowfax\HasEventDispatcher` trait
-- Add `HuangYi\Shadowfax\Events\AppRecyclingEvent`, it will be dispatched when recycling app instance
+- Add `HuangYi\Shadowfax\Events\AppPushingEvent`, it will be dispatched when recycling app instance
 
 
 ## v2.5.1